refactor(onboarding): one reading-reminder component and one state hook
Addresses Ido's two review comments. The step had grown a second component and
a second copy of the state: `useReadingReminder.ts` was a verbatim extraction of
`ReadingReminder.tsx`'s internals, and `OnboardingReadingReminder.tsx` rendered
the same screen with a different layout. Both copies ran the same mount effect,
which is how `/helloworld` ended up logging the step's impression twice.

`ReadingReminder` now takes the state as a prop and renders either layout behind
`isOnboarding`; `useReadingReminder` is the only place that state lives, and the
step calls it once for both arms. `OnboardingReadingReminder` is deleted.

The paid funnel's markup and copy are unchanged — it just receives its state
from above instead of building its own.

Mockup-to-eng-pass: 1
